+++ packages/gafl-webapp-service/src/processors/__tests__/price-display.spec.js
@@ -1,4 +1,4 @@
-import { displayPermissionPrice } from '../price-display.js'
+import { displayPermissionPrice, displayPrice } from '../price-display.js'
 import { getPermissionCost } from '@defra-fish/business-rules-lib'
 
 jest.mock('@defra-fish/business-rules-lib', () => ({
@@ -31,8 +31,19 @@ describe('displayPermissionPrice', () => {
     [48, '#48', getSampleLabels({ pound: '#' })],
     [0, 'Free', getSampleLabels()],
     [0, 'Gratis', getSampleLabels({ free: 'Gratis' })]
-  ])('converts cost returned from getPermissionCost (%d) to string value (%s)', (cost, expectedDisplayPrice, labels) => {
-    getPermissionCost.mockReturnValueOnce(cost)
+  ])('converts cost returned from getPermissionCost (%d) to string value (%s)', (price, expectedDisplayPrice, labels) => {
+    getPermissionCost.mockReturnValueOnce(price)
     expect(displayPermissionPrice({}, labels)).toBe(expectedDisplayPrice)
+  })
+
+  it.each([
+    [27.46, '£27.46', getSampleLabels()],
+    [37.5, '£37.50', getSampleLabels()],
+    [46, '£46', getSampleLabels()],
+    [48, '#48', getSampleLabels({ pound: '#' })],
+    [0, 'Free', getSampleLabels()],
+    [0, 'Gratis', getSampleLabels({ free: 'Gratis' })]
+  ])('displays price %d as %s', (price, expectedDisplayPrice, labels) => {
+    expect(displayPrice(price, labels)).toBe(expectedDisplayPrice)
   })
 })