+++ packages/gafl-webapp-service/src/processors/price-display.js
@@ -1,11 +1,13 @@
 import { getPermissionCost } from '@defra-fish/business-rules-lib'
 
-export const displayPermissionPrice = (permission, labels, createdDate) => {
-  const cost = getPermissionCost(permission, createdDate)
-  if (cost === 0) {
+export const displayPrice = (price, labels) => {
+  if (price === 0) {
     return labels.free
-  } else if (Number.isInteger(cost)) {
-    return `${labels.pound}${cost}`
+  } else if (Number.isInteger(price)) {
+    return `${labels.pound}${price}`
+  } else {
+    return `${labels.pound}${price.toFixed(2)}`
   }
-  return `${labels.pound}${cost.toFixed(2)}`
 }
+
+export const displayPermissionPrice = (permission, labels, createdDate) => displayPrice(getPermissionCost(permission, createdDate), labels)